Refactor ReflectionXml(Json)FormatWriter.
