+++ src/System.Private.DataContractSerialization/src/System/Runtime/Serialization/ReflectionClassWriter.cs
@@ -1,0 +1,156 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Reflection;
+using System.Runtime.Serialization.Json;
+using System.Text;
+using System.Threading.Tasks;
+using System.Xml;
+
+namespace System.Runtime.Serialization
+{
+    internal abstract class ReflectionClassWriter
+    {
+        public void ReflectionWriteClass(XmlWriterDelegator xmlWriter, object obj, XmlObjectSerializerWriteContext context, ClassDataContract classContract, XmlDictionaryString[] memberNames)
+        {
+            InvokeOnSerializing(obj, context, classContract);
+            obj = ResolveAdapterType(obj, classContract);
+            ReflectionWriteMembers(xmlWriter, obj, context, classContract, classContract, 0 /*childElementIndex*/, memberNames);
+            InvokeOnSerialized(obj, context, classContract);
+        }
+
+        public void ReflectionWriteValue(XmlWriterDelegator xmlWriter, XmlObjectSerializerWriteContext context, Type type, object value, bool writeXsiType, PrimitiveDataContract primitiveContractForParamType)
+        {
+            Type memberType = type;
+            object memberValue = value;
+            TypeInfo memberTypeInfo = memberType.GetTypeInfo();
+            bool originValueIsNullableOfT = (memberTypeInfo.IsGenericType && memberType.GetGenericTypeDefinition() == Globals.TypeOfNullable);
+            if (memberTypeInfo.IsValueType && !originValueIsNullableOfT)
+            {
+                PrimitiveDataContract primitiveContract = primitiveContractForParamType;
+                if (primitiveContract != null && !writeXsiType)
+                {
+                    primitiveContract.WriteXmlValue(xmlWriter, memberValue, context);
+                }
+                else
+                {
+                    ReflectionInternalSerialize(xmlWriter, context, memberValue, memberValue.GetType().TypeHandle.Equals(memberType.TypeHandle), writeXsiType, memberType);
+                }
+            }
+            else
+            {
+                if (originValueIsNullableOfT)
+                {
+                    if (memberValue == null)
+                    {
+                        memberType = Nullable.GetUnderlyingType(memberType);
+                    }
+                    else
+                    {
+                        MethodInfo getValue = memberType.GetMethod("get_Value", Array.Empty<Type>());
+                        memberValue = getValue.Invoke(memberValue, Array.Empty<object>());
+                        memberType = memberValue.GetType();
+                    }
+                }
+
+                if (memberValue == null)
+                {
+                    context.WriteNull(xmlWriter, memberType, DataContract.IsTypeSerializable(memberType));
+                }
+                else
+                {
+                    PrimitiveDataContract primitiveContract = originValueIsNullableOfT ? PrimitiveDataContract.GetPrimitiveDataContract(memberType) : primitiveContractForParamType;
+                    if (primitiveContract != null && primitiveContract.UnderlyingType != Globals.TypeOfObject && !writeXsiType)
+                    {
+                        primitiveContract.WriteXmlValue(xmlWriter, memberValue, context);
+                    }
+                    else
+                    {
+                        if (memberValue == null &&
+                            (memberType == Globals.TypeOfObject
+                            || (originValueIsNullableOfT && memberType.GetTypeInfo().IsValueType)))
+                        {
+                            context.WriteNull(xmlWriter, memberType, DataContract.IsTypeSerializable(memberType));
+                        }
+                        else
+                        {
+                            ReflectionInternalSerialize(xmlWriter, context, memberValue, memberValue.GetType().TypeHandle.Equals(memberType.TypeHandle), writeXsiType, memberType, originValueIsNullableOfT);
+                        }
+                    }
+                }
+            }
+        }
+
+        protected abstract int ReflectionWriteMembers(XmlWriterDelegator xmlWriter, object obj, XmlObjectSerializerWriteContext context, ClassDataContract classContract, ClassDataContract derivedMostClassContract, int childElementIndex, XmlDictionaryString[] memberNames);
+
+        protected object ReflectionGetMemberValue(object obj, DataMember dataMember)
+        {
+            return dataMember.Getter(obj);
+        }
+
+        protected bool ReflectionTryWritePrimitive(XmlWriterDelegator xmlWriter, XmlObjectSerializerWriteContext context, Type type, object value, XmlDictionaryString name, XmlDictionaryString ns, PrimitiveDataContract primitiveContract)
+        {
+            if (primitiveContract == null || primitiveContract.UnderlyingType == Globals.TypeOfObject)
+                return false;
+
+            primitiveContract.WriteXmlElement(xmlWriter, value, context, name, ns);
+
+            return true;
+        } 
+
+        private void InvokeOnSerializing(object obj, XmlObjectSerializerWriteContext context, ClassDataContract classContract)
+        {
+            if (classContract.BaseContract != null)
+                InvokeOnSerializing(obj, context, classContract.BaseContract);
+            if (classContract.OnSerializing != null)
+            {
+                var contextArg = context.GetStreamingContext();
+                classContract.OnSerializing.Invoke(obj, new object[] { contextArg });
+            }
+        }
+
+        private void InvokeOnSerialized(object obj, XmlObjectSerializerWriteContext context, ClassDataContract classContract)
+        {
+            if (classContract.BaseContract != null)
+                InvokeOnSerialized(obj, context, classContract.BaseContract);
+            if (classContract.OnSerialized != null)
+            {
+                var contextArg = context.GetStreamingContext();
+                classContract.OnSerialized.Invoke(obj, new object[] { contextArg });
+            }
+        }
+
+        private object ResolveAdapterType(object obj, ClassDataContract classContract)
+        {
+            Type type = obj.GetType();
+            if (type == Globals.TypeOfDateTimeOffset)
+            {
+                obj = DateTimeOffsetAdapter.GetDateTimeOffsetAdapter((DateTimeOffset)obj);
+            }
+            else if (type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == Globals.TypeOfKeyValuePair)
+            {
+                obj = classContract.KeyValuePairAdapterConstructorInfo.Invoke(new object[] { obj });
+            }
+
+            return obj;
+        }
+
+        private void ReflectionInternalSerialize(XmlWriterDelegator xmlWriter, XmlObjectSerializerWriteContext context, object obj, bool isDeclaredType, bool writeXsiType, Type memberType, bool isNullableOfT = false)
+        {
+            if (isNullableOfT)
+            {
+                context.InternalSerialize(xmlWriter, obj, isDeclaredType, writeXsiType, DataContract.GetId(memberType.TypeHandle), memberType.TypeHandle);
+            }
+            else
+            {
+                context.InternalSerializeReference(xmlWriter, obj, isDeclaredType, writeXsiType, DataContract.GetId(memberType.TypeHandle), memberType.TypeHandle);
+            }
+        }
+    }
+}
