+++ src/System.Runtime.Serialization.Xml/tests/SerializationTypes.cs
@@ -2403,13 +2403,13 @@ namespace SerializationTypes
         public SimpleType[][] TwoDArrayOfSimpleType;
     }
 
-    public class SimpleTypeWihtMoreFields
+    public class SimpleTypeWithMoreFields
     {
         public string StringField;
         public int IntField;
         public MyEnum EnumField;
         public List<string> CollectionField;
-        public List<SimpleTypeWihtMoreFields> SimpleTypeList;
+        public List<SimpleTypeWithMoreFields> SimpleTypeList;
     }
 
     // New types
